+++ scilab/modules/renderer/src/cpp/subwinDrawing/BackTrihedronDrawerJoGL.cpp
@@ -40,7 +40,7 @@ void BackTrihedronDrawerJoGL::drawAxesBox(int concealedCornerIndex)
   initializeDrawing();
 
   getBackTrihedronDrawerJavaMapper()->setBoxParameters(pSUBWIN_FEATURE(pSubwin)->axes.hiddenAxisColor,
-                                                       sciGetLineWidth(pSubwin));
+                                                       (float)sciGetLineWidth(pSubwin));
 
   // get displayed bounds
   double bounds[6];
